refactor: organize component structure
