+++ src/components/sections/AppSection.tsx
@@ -1,5 +1,5 @@
-import { Card } from "@/components/ui/card"
-import { Button } from "@/components/ui/button"
+import { Card } from "@/components/ui/Card"
+import { Button } from "@/components/ui/Button"
 
 interface AppSectionItem {
   id: number
