+++ src/components/sections/AvatarGrid.tsx
